+++ src/components/icon-button/icon-button.styles.ts
@@ -42,6 +42,6 @@ export default css`
   }
 
   .focus-visible.icon-button:focus {
-    box-shadow: 0 0 0 var(--sl-focus-ring-width) rgb(var(--sl-focus-ring-color-primary) / var(--sl-focus-ring-alpha));
+    box-shadow: 0 0 0 var(--sl-focus-ring-width) rgb(var(--sl-color-primary-500) / var(--sl-focus-ring-alpha));
   }
 `;