+++ src/components/rating/rating.styles.ts
@@ -25,7 +25,7 @@ export default css`
   }
 
   .rating.focus-visible:focus {
-    box-shadow: 0 0 0 var(--sl-focus-ring-width) rgb(var(--sl-focus-ring-color-primary) / var(--sl-focus-ring-alpha));
+    box-shadow: 0 0 0 var(--sl-focus-ring-width) rgb(var(--sl-color-primary-500) / var(--sl-focus-ring-alpha));
   }
 
   .rating__symbols {