+++ src/components/tab/tab.styles.ts
@@ -5,8 +5,7 @@ export default css`
   ${componentStyles}
 
   :host {
-    --focus-ring: inset 0 0 0 var(--sl-focus-ring-width)
-      rgb(var(--sl-focus-ring-color-primary) / var(--sl-focus-ring-alpha));
+    --focus-ring: inset 0 0 0 var(--sl-focus-ring-width) rgb(var(--sl-color-primary-500) / var(--sl-focus-ring-alpha));
 
     display: inline-block;
   }